requirements.txt
fastapi
uvicorn
python-jose[cryptography]
bcrypt
sendgrid
pillow
sqlalchemy
python-dotenv
cryptography
aiofiles
pydantic[email]
python-multipart